sc-executor-wasmi: introduce 'interpreted-wasm' feature
