+++ client/executor/Cargo.toml
@@ -17,7 +17,6 @@ targets = ["x86_64-unknown-linux-gnu"]
 lru = "0.8.1"
 parking_lot = "0.12.1"
 tracing = "0.1.29"
-wasmi = "0.13.2"
 
 codec = { package = "parity-scale-codec", version = "3.2.2" }
 sc-executor-common = { version = "0.10.0-dev", path = "common" }
@@ -61,3 +60,4 @@ default = ["std"]
 # This crate does not have `no_std` support, we just require this for tests
 std = []
 wasm-extern-trace = []
+interpreted-wasm = ["sc-executor-wasmi/enabled"]
