+++ client/executor/wasmi/Cargo.toml
@@ -15,8 +15,10 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 log = "0.4.17"
-wasmi = { version = "0.13.2", features = [ "virtual_memory" ] }
 sc-allocator = { version = "4.1.0-dev", path = "../../allocator" }
 sc-executor-common = { version = "0.10.0-dev", path = "../common" }
 sp-runtime-interface = { version = "7.0.0", path = "../../../primitives/runtime-interface" }
 sp-wasm-interface = { version = "7.0.0", path = "../../../primitives/wasm-interface" }
+
+[features]
+enabled = ["sp-wasm-interface/wasmi", "sc-executor-common/wasmi"]
