+++ primitives/wasm-interface/Cargo.toml
@@ -17,12 +17,12 @@ targets = ["x86_64-unknown-linux-gnu"]
 codec = { package = "parity-scale-codec", version = "3.2.2", default-features = false, features = ["derive"] }
 impl-trait-for-tuples = "0.2.2"
 log = { version = "0.4.17", optional = true }
-wasmi = { version = "0.13.2", optional = true }
+wasmi = { version = "0.13.2", features = [ "virtual_memory" ], optional = true }
 wasmtime = { version = "6.0.1", default-features = false, optional = true }
 anyhow = { version = "1.0.68", optional = true }
 sp-std = { version = "5.0.0", default-features = false, path = "../std" }
 
 [features]
 default = [ "std" ]
-std = [ "codec/std", "log", "sp-std/std", "wasmi", "wasmtime" ]
+std = [ "codec/std", "log", "sp-std/std", "wasmtime" ]
 wasmtime = [ "dep:wasmtime", "anyhow" ]
