+++ primitives/wasm-interface/src/lib.rs
@@ -22,6 +22,9 @@
 use sp_std::{borrow::Cow, iter::Iterator, marker::PhantomData, mem, result, vec, vec::Vec};
 
 #[cfg(feature = "wasmi")]
+pub use wasmi;
+
+#[cfg(feature = "wasmi")]
 mod wasmi_impl;
 
 #[cfg(not(all(feature = "std", feature = "wasmtime")))]
